Add files via upload
Now my algorithm implemented in multiple functions finds all the needed dot cycles. Regardless of the fact whether they are connected or not. Regardless of the fact how many of them are out there. Now is the time to work on finding if there are any enemy dots inside these cycles.  And to think how to write a color filling function