Add files via upload
Now opacity is a factor in the game graphics. The camera zooms to the center, not to the left top angle. The scaling and the coordinate system are improved a bit. The dots appear where they should, no more bugs, when you click to draw a dot and it simply does not appear, so you should zoom in and out to be able to draw it. The bug when the Select amount of players menu suddenly throws you back to the main menu has been fixed.