Add files via upload
Minor changes in scaling the mouse arrow and dot scaling. The Zoom() function is now shorter and more simple. Bug fix with changing the fourth player color in FourPlayerMenu() with keyboard. Background field theme changed. Now the numbers of columns and rows change their color along with the background theme in the custom menu.